Debugging: fw_dump: Save `devlink health` output
Upon a firmware fatal event, in addition to triggering a CR space dump,
also retrieve the last saved dump from the kernel. It contains
information about the event trigger. For example:

$ devlink health dump show DEV reporter fw_fatal
 irisc_id: 0
  event:
    id: 1 desc: CR space timeout
  method: write long_process: false command_type: emad reg_attr_id:37104 log_address: 12592896 log_irisc_id: 1

From the output it is possible to know the latest used register,
accessed address and more.

Save this information as part of the tar file, in file named
devlink_health_dump.txt.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>