libbpf-tools: Add a top-level build system
The following commits will add resmon, a tool for tracking resource usage
on a Spectrum switch. resmon consists of a daemon, a CLI interface, a BPF
program, and is more complex than the other libbpf-based tools. It will
therefore be in a separete directory, with its own Makefile. However it
will still benefit from using an up-to-date libbpf, and should generally
reuse the libbpf-related infrastructure.

In this patch, move the libbpf-building logic to a new Makefile fragment,
common.mk. Also add a top-level Makefile for dispatching common targets to
individual sub-directories.

Signed-off-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>