resmon: Allow to select which resources to monitor
Currently, all supported resources are monitored when resmon is started.

For case that user is interested only in certain resources, add the ability
to filter which resources to count.

To start resmon and choose only several resources, run for example:
$ ./resmon start resources lpm_ipv6

In addition, similar resources can be grouped:
To filter lpm_ipv4 and lpm_ipv6, user can use "lpm".

Ignore registers that are not related to the selected resources and print
only the relevant gauges.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>
Co-developed-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>