resmon: Support PTAR / PTCE3
TCAM regions are allocated through the PTAR register. PTAR.tcam_region_info
in the response EMAD holds the unique key of the allocated or freed
resource. Number of KVD slots to consume is determined from
PTAR.flexible_keys in allocation requests. In deletion requests, no keys
are passed, and thus we need to rely solely on the KVDA in the PTAR map.

Signed-off-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>