resmon: Support SFD
FDB entries are maintained through SFD register.

There are several records types for allocating/deallocating entries.
The keys which are common to all the types are MAC and FID and each record
type contains a parameter (for example, system_port for unicast, lag_id
for unicast lag, etc..).

FDB entries can be flushed by SFDF register, according to several
parameters, for example, it is possible to flush all entries with FID=X
which point to system_port=Y.

Save a dedicated hash table for FDB entries to be able to flush the entries
in the next patch.

Table sfd saves param_type and param as part of the value:
(mac,fid)->(kvd_alloc,param_type,param)

The payload of SFD register is long, so it should be possible to read a
longer length than the current limitation, otherwise, push_to_ringbuf()
will ignore these EMADs.

$ ./resmon/resmon stats
Resource                      Usage
...
FDB Entry                     0 / 524288 (0%)

Signed-off-by: Amit Cohen <amcohen@nvidia.com>
Signed-off-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>