resmon: Add a CLI option to show result in JSON format
Currently when one wishes to parse resmon output, they have to resort to
coding the low-level JSON RPC. The testsuite in resmon itself uses that
approach, proving it is a relevant use case.

To better support this use case, add a command-line option, --json. When
specified, successful result of the RPC is not interpreted and formatted in
a human-readable way, but instead dumped as a JSON object.

Convert the test suite to use this interface instead of hand-rolling the
RPC through netcat.

Signed-off-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>