resmon: Rename two functions
The function resmon_c_handle_response() actually doesn't "handle" the
response per se. It dissects and validates it, and when it is an error
response, prints the error. The goal actually is to get the response
result, if present, and bail out if not. Rename it appropriately, to
resmon_c_response_extract_result(). This name also respects the established
subject-verb-object approach to naming identifiers.

For consistency, also rename resmon_c_handle_response_error() to
resmon_c_response_handle_error().

Signed-off-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>