resmon: Extract resmon_c_send_request_on()
In order to dump a table, a number of requests need to be made to dump the
individual rows of the table. Creating the Unix socket for each request
separately is inefficient. Therefore extract from resmon_c_send_request() a
helper that does the transport and response parsing itself, so that it can
later be reused from the dump handler.

Signed-off-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>