resmon: Add a table abstraction
In order to keep track of dumping, it will be useful to promote resmon
tables into a recognized abstraction. This will allow a centralized way of
storing per-table metadata.

There are several helpers that were meant to act on specifically resmon
tables before: the resmon_stat_lh_ family of functions. Update their
interface to take table pointer instead of the raw LH pointer. Update the
name appropriately as well.

All the tables have used and do use predictably-named functions for key
hashing and comparison. Since we are touching this code anyway, extract the
commonality into a macro that just autogenerates the proper table
initialization form, thereby reducing the risk of typos and discrepancies
in used eq and hash functions.

Signed-off-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>