resmon: Add dump_next RPC
The dump_next RPC will serve for dumping of a given table. This works my
calling the dump_next RPC repeatedly, each time getting a new row.

This is a very simple interface meant mostly for debugging, rather than
DB-like querying. The dump can be inconsistent if there were changes to the
table during the dumping, or if there is a concurrent second dumper, the
dump might be incomplete, but actually might end up duplicating rows
instead. Two heuristics help determine whether the dump was consistent: if
the number of rows dumped matches the number of rows that the table claimed
to have before the dump, that is an indication that the client has seen all
the rows. And if the sequence number after the dump and before the dump is
the same, that is an indication that no rows were added or dropped. Neither
of these is surefire, but for what the interface is, they probably serve
just fine.

Client-side, the responses (if any) are either dumped directly as JSON, in
--json mode, or are formatted in a human-readable fashion in a table.

Signed-off-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>