Add a tool for computing headroom buffer size
When configuring a switch for lossless traffic, it is necessary to give
headroom buffers enough size to accomodate traffic that is still to arrive
after the need to pause the traffic is identified. Determining the buffer
size is not trivial and entails a number of unobvious considerations.

Add a tool to assist users with this calculation.

Signed-off-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>