resmon: Increase size of BPF ring buffer
Increase in the number of entries in a hash table used to store KVD
entries (e.g., routes) can result in the hash table being resized. This
incurs a performance penalty [1] which can lead to EMADs being buffered
in the BPF ring buffer.

If the resizing takes too long and EMADs are processed at a high rate,
it is possible for the ring buffer to become full, resulting in EMADs
being dropped and resmon displaying incorrect statistics.

On SN4700 we see the following results when pushing 256K IPv4 routes:

Ring buffer size | Missing Entries
-----------------|----------------
256 KB           | 1706
512 KB           | 388
1 MB             | 0

Fix the problem by increasing the ring buffer size to 2 MB, taking into
account both larger number of entries and larger EMADs. Tested
successfully with 458,752 IPv4 routes.

Another option is to increase the initial size of each hash table, but
this is wasteful as not all the hash tables can be full at the same
time.

[1] https://json-c.github.io/json-c/json-c-0.10/doc/html/linkhash_8h.html#ac3ba464ac43c083397b1f1d34f2ddf55

Signed-off-by: Ido Schimmel <idosch@nvidia.com>
Reviewed-by: Petr Machata <petrm@nvidia.com>