resmon: Increase initial hash table size
On some systems, the recent increase in ring buffer size is not enough
and EMADs are dropped by the BPF program upon hash table resizing.

Increasing the initial hash table size to 10 seems to solve the problem.
This number is not indicative of the entries we are expected to have in
the hash table, but it works as a bigger initial size results in fewer
resizing operations [1].

Fix the problem by increasing the initial hash table size to 1000. This
results in 2.84% increase in initial memory usage.

[1] https://github.com/json-c/json-c/blob/master/linkhash.c#L586

Reported-by: Maksym Yaremchuk <maksymy@nvidia.com>
Tested-by: Maksym Yaremchuk <maksymy@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>
Reviewed-by: Petr Machata <petrm@nvidia.com>