Debugging: drgn-tools: Duplicate 'fid_dump' tool for old kernels
The kernel commit 798661c73672 ("Merge branch 'mlxsw-unified-bridge-conversion-part-6'")
merged the last patches which are required to support unified bridge model.

The next patches will adjust 'fid_dump' tool to support new fields which
are used with the new model. To allow debugging old kernels, add a copy
of the existing tool and call it 'fid_dump_ub0'.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>
Reviewed-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>