Debugging: drgn-tools: fid_dump: Dump MID indexes per FID
As part of converting mlxsw driver to use unified bridge model, a
'mid_index' is calculated for flooding unicast, multicast and broadcast
packets. These indexes are the indexes to PGT table and are the result
of 'mid_base' + 'fid_offset'. See kernel commit 9f6f467a3cdb1 ("mlxsw:
spectrum_fid: Set 'mid_base' as part of flood tables initialization")

The MID indexes are not stored as part of FID structure, calculate them
in the script in order to dump them. These values are not relevant for
RFIDs and dummy FID, so calculate them only for 802.1Q and 802.1D FIDs.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>
Reviewed-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>