libbpf-tools: emadlatency.bpf.c: Add String TLV and Latency TLV fields
The following patches will use String TLV and Latency TLV from EMAD's
header to get the latency time which is measured by firmware.

Add some useful macros, structures and helpers to recognize these TLVs.
Note that String TLV and Latency TLV are optional, so the program will
have to verify the type of the TLV to check if these TLVs are attached to
the EMAD or not. Increase 'EMAD_HDR_LEN' which defines how many bytes will
be read for each EMAD, so then the program will read String and Latency
TLVs.

Signed-off-by: Danielle Ratson <danieller@nvidia.com>
Signed-off-by: Amit Cohen <amcohen@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>