libbpf-tools: emadlatency.bpf.c: Read and save firmware latency measurements
The latency of each EMAD is measured by firmware. The driver can get the
measurement via Latency TLV which can be added to each EMAD.

The structure of EMAD is:
* Ethernet header
* Operation TLV
* String TLV (optional)
* Latency TLV (optional)
* Reg TLV
* End TLV

Till now, the BPF program reads only the first two parts in EMAD's
header. The latency time which is measured by firmware is part of Latency
TLV, which is an optional TLV.

Add support for reading 'latency-time' field from Latency TLV and insert
the values to a new histogram for firmware measurements. The next patches
will print these measurements in user-space code.

Note that both String TLV and Latency TLV are optional, so the program
first has to check if these TLVs are attached to the EMAD, in order to
find the correct bytes which represent the 'latency-time' field. Each
TLV contains a type, which can be used to identify which TLV is read.

Signed-off-by: Danielle Ratson <danieller@nvidia.com>
Signed-off-by: Amit Cohen <amcohen@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>