resmon: Makefile: Move objects before libraries they depend on
The Ubuntu 22 toolchain injects --as-needed to the linker command line.
This option causes the resulting object to only link with any given library
if at the point where that library is mentioned, it is actually necessary
for resolution of any symbols. Since we put the libraries before the
objects that use them, where no symbols need to be resolved yet, the
libraries are not pulled in, and the resulting link fails with an array of
unresolved symbol citations.

This issue can be reproduced by passing the option explicitly:

 # make CC="gcc -Wl,--as-needed"

Fix by moving the objects to the beginning of the command line.

Signed-off-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>