resmon: stat: Zero the allocated resmon_stat in resmon_stat_create()
The field struct resmon_stat::gauges seems to happen to consistently start
with zeroes, but valgrind complains about uninitialized memory when
formatting the stats RPC, and I suspect that MALLOC_PERTURB_ would cause it
to break as well.

Most of the time we initialize structures using the *x = (struct x){...}
idiom, so allocating an all-zero memory is not necessary. Here, however, we
do a field-by-field allocation, which is why neglecting to include a field
caused this issue. So to fix this problem and future-proof this part of
code, request a zeroed-out memory by allocating through calloc().

Signed-off-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>