resmon: bpf: Fix missing EMAD responses
We often see resmon complaining that some entries (e.g., routes) were
not deleted after certain tests. Petr debugged it down to the fact that
the resmon BPF program is not invoked for every EMAD response.

The resmon BPF program is attached to the 'devlink:devlink_hwmsg'
tracepoint which is triggered by mlxsw upon EMAD transmission and
reception. The former happens in process context while the latter
happens in softIRQ context.

It is therefore possible that a softIRQ is raised while the BPF program
is processing an EMAD request. If the softIRQ is raised because of an
incoming EMAD response, then the kernel will try to invoke the BPF
program on the same CPU. According to [1], this will result in the BPF
program not being invoked for the EMAD response:

"kprobe programs have per-CPU reentrancy protection. That is, if some
BPF kprobe/tracepoint program is running and something happens [...]
that would trigger another BPF program, then that nested BPF program
invocation will be skipped."

Fix this by attaching the BPF program to the mlxsw function that
processes EMAD responses. This function is always executed in softIRQ
context so it is not possible to run into the above problem.

With this fix the problem did not reproduce in 600 iterations on SN2201,
whereas without the fix the problem was reproduced quite consistently.

[1] https://lore.kernel.org/all/CAEf4BzZg0r4YptYPu8Y_-qp=rY__W6dmb9kLwMV5MAH6C-2PSg@mail.gmail.com/

Fixes: e7ca40f02eb4 ("resmon: Introduce the BPF interface")
Signed-off-by: Ido Schimmel <idosch@nvidia.com>
Reviewed-by: Petr Machata <petrm@nvidia.com>