resmon: Increase size of BPF ring buffer once again
When adding 256K FDB entries on SN4700 we see a mismatch in the number
of configured FDB entries between the bridge driver and resmon. The
reason is that the resmon user space process does not receive all the
SFD EMADs from the kernel due to insufficient space in the BPF ring
buffer as evident by inspecting the BPF trace buffer:

 # cat /sys/kernel/debug/tracing/trace_pipe
  napi/mlxsw_rx-0-5281    [001] b..11  3245.528181: bpf_trace_printk: Unable to reserve 1224 bytes for an EMAD on ring buffer

  napi/mlxsw_rx-0-5281    [001] b..11  3245.528233: bpf_trace_printk: Unable to reserve 1224 bytes for an EMAD on ring buffer
  [...]

Fix by increasing the size of the ring buffer once again in a similar
fashion to commit a7b4ced6b358 ("resmon: Increase size of BPF ring
buffer"). With the new size we no longer miss events from the kernel.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>